functions/aws/aws-image-resizer/requirements.txt
pillow:python3.9
aws_lambda_powertools